made some changes to the file structure
