--- src/main/deploy/paths.json
@@ -1,5 +1,0 @@
-[
-
-
-
-]
