+++ src/main/deploy/paths/TakeRingAndShoot.json
@@ -1,0 +1,7 @@
+[
+  "shootSpeaker",
+  ["gotoClosestRing","intakeDown"],
+  "intakeUp",
+  ["gotoClosestSpeaker"],
+  "shootSpeaker"
+]
