+++ src/main/kotlin/frc/robot/Robot.kt
@@ -1,12 +1,8 @@
 package frc.robot
 
-import edu.wpi.first.math.geometry.Pose2d
-import edu.wpi.first.wpilibj.DriverStation
 import edu.wpi.first.wpilibj.TimedRobot
 import edu.wpi.first.wpilibj2.command.Command
 import edu.wpi.first.wpilibj2.command.CommandScheduler
-import edu.wpi.first.wpilibj2.command.Commands.runOnce
-import frc.lib.constants.SwerveDriveConstants
 
 /**
  * The VM is configured to automatically run this class, and to call the functions corresponding to
